优化ToAMConverter
