+++ opscloud-manage/src/main/java/com/baiyi/opscloud/packer/user/am/converter/AbstractToAMConverter.java
@@ -1,6 +1,10 @@
 package com.baiyi.opscloud.packer.user.am.converter;
 
 import com.baiyi.opscloud.core.factory.DsConfigHelper;
+import com.baiyi.opscloud.domain.generator.opscloud.DatasourceConfig;
+import com.baiyi.opscloud.domain.generator.opscloud.DatasourceInstance;
+import com.baiyi.opscloud.domain.vo.datasource.DsAssetVO;
+import com.baiyi.opscloud.domain.vo.user.AMVO;
 import com.baiyi.opscloud.packer.user.am.AmPacker;
 import com.baiyi.opscloud.packer.user.am.IToAMConverter;
 import com.baiyi.opscloud.service.datasource.DsConfigService;
@@ -8,6 +12,7 @@ import com.baiyi.opscloud.service.datasource.DsInstanceService;
 import org.springframework.beans.factory.InitializingBean;
 
 import javax.annotation.Resource;
+import java.util.List;
 
 /**
  * @Author baiyi
@@ -24,6 +29,29 @@ public abstract class AbstractToAMConverter implements IToAMConverter, Initializ
 
     @Resource
     protected DsConfigHelper dsConfigHelper;
+
+    @Override
+    public AMVO.XAM toAM(DsAssetVO.Asset asset) {
+        DatasourceInstance instance = dsInstanceService.getByUuid(asset.getInstanceUuid());
+        DatasourceConfig datasourceConfig = dsConfigService.getById(instance.getConfigId());
+        List<DsAssetVO.Asset> policies = toPolicies(asset);
+        AMVO.XAM xam = AMVO.XAM.builder()
+                .instanceUuid(instance.getUuid())
+                .instanceName(instance.getInstanceName())
+                .username(asset.getAssetKey())
+                .name(asset.getName())
+                .accessKeys(toAccessKeys(asset))
+                .policies(toPolicies(asset))
+                .build();
+        wrap(xam,datasourceConfig);
+        return xam;
+    }
+
+    abstract protected void wrap(AMVO.XAM xam,DatasourceConfig datasourceConfig);
+
+    abstract protected List<DsAssetVO.Asset> toAccessKeys(DsAssetVO.Asset asset);
+
+    abstract protected List<DsAssetVO.Asset> toPolicies(DsAssetVO.Asset asset);
 
     @Override
     public void afterPropertiesSet() throws Exception {
